Verify docker service configuration.
This commit adds check-docker-config.sh, which ensures the cgroup is set in /etc/sysconfig/docker. It does not check for the other options (devicemapper and dns/bip) as those are checked in other scripts (docker-storage.sh and check-bip-dns.py, respectively).

The commit also modifies the check-docker-service.py script:

* Don't bail on one failure, if there are multiple failures in the service file, they are all shown
* Now checks for correct ExecStart and TasksMax

Implements INSP-2